+++ routes/index.mjs
@@ -17,7 +17,7 @@ import config from '../services/config';
 import ratings from './ratings';
 import stats from './stats';
 import shuttles from './shuttles';
-import patterns from './patterns';
+import shuttleFactory from './shuttle-factories';
 
 const router = new Router();
 
@@ -59,7 +59,7 @@ router.use('/stats', stats.routes());
 router.use('/stats', stats.allowedMethods());
 router.use('/shuttles', shuttles.routes());
 router.use('/shuttles', shuttles.allowedMethods());
-router.use('/patterns', patterns.routes());
-router.use('/patterns', patterns.allowedMethods());
+router.use('/shuttle-factories', shuttleFactory.routes());
+router.use('/shuttle-factories', shuttleFactory.allowedMethods());
 
 export default router;