+++ tests/models/pattern.mjs
@@ -1,7 +1,0 @@
-import Pattern from '../../models/pattern';
-
-export const generateDummyPattern = (params) => ({ ...params });
-
-export const createDummyPattern = async (params) => Pattern.create(generateDummyPattern(params));
-
-export default Pattern;