+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 * Added capacity check in ride pre validate hook
 * Changed logger from Winston to Pino
 * Added shuttle routes and model
-* Added pattern routes and model
+* Added shuttle factory routes and model
 ### Fix
 ## Version 1.4.6c
 ### Fix